pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "muto_link"
version = "0.1.0"
description = "Hardware-agnostic serial link with pluggable transports (USB, Pi UART/GPIO)."
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["typer>=0.12.0", "pyserial>=3.5"]

# Optional Pi extras; install with: uv sync -E pi  (or uv pip install ".[pi]")
[project.optional-dependencies]
pi = ["gpiozero>=2.0"]
dev = ["pytest>=6.0"]

[dependency-groups]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.24.0",
    "mkdocs-gen-files>=0.5.0",
    "mkdocs-literate-nav>=0.6.0",
    "mkdocs-git-revision-date-localized-plugin>=1.2.0",
]

[project.scripts]
muto = "app.cli:app"

[tool.uv]
package = true

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
addopts = "-q --disable-warnings"
